RecoveryRecord: add date fields
These fields are intentionally not of type time.Time because they only contain a
date, not a timestamp. The time zone is not specified, but in practice can be
assumed to be local to the issuing country. We can consider adding an accessor
in the future which attempts to parse the date, but for now the fields just
contain what the QR code contains.

fixes https://github.com/stapelberg/coronaqr/issues/11